ex4

+++ ex3/displayData.py
@@ -12,12 +12,12 @@ def displayData(X):
 
 # Compute rows, cols
     m, n = X.shape
-    example_width = round(np.sqrt(n))
+    example_width = int(round(np.sqrt(n)))
     example_height = (n / example_width)
 
 # Compute number of items to display
-    display_rows = np.floor(np.sqrt(m))
-    display_cols = np.ceil(m / display_rows)
+    display_rows = np.floor(np.sqrt(m)).astype(np.int)
+    display_cols = np.ceil(m / display_rows).astype(np.int)
 
 # Between images padding
     pad = 1
